add roboteq robot driver abstraction
