+++ panther_hardware_interfaces/src/panther_system/robot_driver/lynx_robot_driver.cpp
@@ -14,230 +14,55 @@
 
 #include "panther_hardware_interfaces/panther_system/robot_driver/lynx_robot_driver.hpp"
 
-#include <chrono>
-#include <ctime>
 #include <memory>
 #include <stdexcept>
 #include <string>
-#include <thread>
-#include <utility>
+#include <vector>
 
-#include "lely/util/chrono.hpp"
-
-#include "panther_hardware_interfaces/panther_system/robot_driver/canopen_manager.hpp"
-#include "panther_hardware_interfaces/panther_system/robot_driver/roboteq_data_converters.hpp"
 #include "panther_hardware_interfaces/panther_system/robot_driver/roboteq_driver.hpp"
+#include "panther_hardware_interfaces/panther_system/robot_driver/roboteq_robot_driver.hpp"
 
 namespace panther_hardware_interfaces
 {
 
-LynxRobotDriver::LynxRobotDriver(
-  const CANopenSettings & canopen_settings, const DrivetrainSettings & drivetrain_settings,
-  const std::chrono::milliseconds activate_wait_time)
-: canopen_settings_(canopen_settings),
-  canopen_manager_(canopen_settings),
-  data_(drivetrain_settings),
-  roboteq_vel_cmd_converter_(drivetrain_settings),
-  pdo_motor_states_timeout_ms_(canopen_settings.pdo_motor_states_timeout_ms),
-  pdo_driver_state_timeout_ms_(canopen_settings.pdo_driver_state_timeout_ms),
-  activate_wait_time_(activate_wait_time)
+void LynxRobotDriver::SendSpeedCommands(const std::vector<float> & speeds)
 {
-}
-
-void LynxRobotDriver::Initialize()
-{
-  if (initialized_) {
-    return;
-  }
-
-  try {
-    canopen_manager_.Initialize();
-    DefineDriver();
-    driver_->Boot();
-  } catch (const std::runtime_error & e) {
-    throw e;
-  }
-
-  initialized_ = true;
-}
-
-void LynxRobotDriver::Deinitialize()
-{
-  canopen_manager_.Deinitialize();
-  initialized_ = false;
-}
-
-void LynxRobotDriver::Activate()
-{
-  // Activation procedure - it is necessary to first reset scripts, wait for a bit (1 second)
-  // and then send 0 commands for some time (also 1 second)
-
-  try {
-    driver_->ResetScript();
-  } catch (const std::runtime_error & e) {
+  if (speeds.size() != 2) {
     throw std::runtime_error(
-      "Front driver reset Roboteq script exception: " + std::string(e.what()));
+      "Invalid speeds vector size. Expected 2, got " + std::to_string(speeds.size()));
   }
 
-  std::this_thread::sleep_for(activate_wait_time_);
+  const auto speed_left = this->GetCmdVelConverter().Convert(speeds.at(0));
+  const auto speed_right = this->GetCmdVelConverter().Convert(speeds.at(1));
 
   try {
-    driver_->GetMotorDriver(LynxMotorNames::LEFT)->SendCmdVel(0);
-    driver_->GetMotorDriver(LynxMotorNames::RIGHT)->SendCmdVel(0);
+    drivers_.at(DriverNames::DEFAULT)->GetMotorDriver(MotorNames::LEFT)->SendCmdVel(speed_left);
+    drivers_.at(DriverNames::DEFAULT)->GetMotorDriver(MotorNames::RIGHT)->SendCmdVel(speed_right);
   } catch (const std::runtime_error & e) {
-    throw std::runtime_error("Front driver send 0 command exception: " + std::string(e.what()));
+    throw std::runtime_error("Driver send Roboteq cmd failed: " + std::string(e.what()));
   }
 
-  std::this_thread::sleep_for(activate_wait_time_);
-}
-
-void LynxRobotDriver::UpdateCommunicationState()
-{
-  data_.SetCANError(driver_->IsCANError());
-  data_.SetHeartbeatTimeout(driver_->IsHeartbeatTimeout());
-}
-
-void LynxRobotDriver::UpdateMotorsState()
-{
-  timespec current_time;
-  clock_gettime(CLOCK_MONOTONIC, &current_time);
-
-  const auto fl_state = driver_->GetMotorDriver(LynxMotorNames::LEFT)->ReadMotorDriverState();
-  const auto fr_state = driver_->GetMotorDriver(LynxMotorNames::RIGHT)->ReadMotorDriverState();
-
-  SetMotorsStates(data_, fl_state, fr_state, current_time);
-
-  UpdateCommunicationState();
-
-  if (data_.IsCANError()) {
-    throw std::runtime_error("CAN error.");
-  }
-
-  if (data_.IsHeartbeatTimeout()) {
-    throw std::runtime_error("Motor controller heartbeat timeout.");
-  }
-}
-
-void LynxRobotDriver::UpdateDriversState()
-{
-  timespec current_time;
-  clock_gettime(CLOCK_MONOTONIC, &current_time);
-
-  SetDriverState(data_, driver_->ReadDriverState(), current_time);
-
-  UpdateCommunicationState();
-
-  if (data_.IsCANError()) {
-    throw std::runtime_error("CAN error.");
-  }
-
-  if (data_.IsHeartbeatTimeout()) {
-    throw std::runtime_error("Motor controller heartbeat timeout.");
-  }
-}
-
-const RoboteqData & LynxRobotDriver::GetData(const std::string & name)
-{
-  if (name == LynxDriverNames::DEFAULT) {
-    return data_;
-  } else {
-    throw std::runtime_error("Data with name '" + name + "' does not exist.");
-  }
-}
-
-void LynxRobotDriver::SendSpeedCommands(
-  const float speed_left, const float speed_right, const float /*speed_rl*/,
-  const float /*speed_rr*/)
-{
-  // Channel 1 - right motor, Channel 2 - left motor
-  try {
-    driver_->GetMotorDriver(LynxMotorNames::LEFT)
-      ->SendCmdVel(roboteq_vel_cmd_converter_.Convert(speed_left));
-    driver_->GetMotorDriver(LynxMotorNames::RIGHT)
-      ->SendCmdVel(roboteq_vel_cmd_converter_.Convert(speed_right));
-  } catch (const std::runtime_error & e) {
-    throw std::runtime_error("Front driver send Roboteq cmd failed: " + std::string(e.what()));
-  }
-
-  if (driver_->IsCANError()) {
+  if (drivers_.at(DriverNames::DEFAULT)->IsCANError()) {
     throw std::runtime_error(
-      "CAN error detected on the front driver when trying to write speed commands.");
+      "CAN error detected on the Driver when trying to write speed commands.");
   }
 }
 
-void LynxRobotDriver::TurnOnEStop()
+void LynxRobotDriver::DefineDrivers()
 {
-  try {
-    driver_->TurnOnEStop();
-  } catch (const std::runtime_error & e) {
-    throw std::runtime_error(
-      "Failed to turn on E-stop on the front driver: " + std::string(e.what()));
-  }
-}
-
-void LynxRobotDriver::TurnOffEStop()
-{
-  try {
-    driver_->TurnOffEStop();
-  } catch (const std::runtime_error & e) {
-    throw std::runtime_error(
-      "Failed to turn off E-stop on the front driver: " + std::string(e.what()));
-  }
-}
-
-void LynxRobotDriver::TurnOnSafetyStop()
-{
-  try {
-    driver_->GetMotorDriver(LynxMotorNames::LEFT)->TurnOnSafetyStop();
-    driver_->GetMotorDriver(LynxMotorNames::RIGHT)->TurnOnSafetyStop();
-  } catch (const std::runtime_error & e) {
-    throw std::runtime_error(
-      "Failed to turn on safety stop on the front driver: " + std::string(e.what()));
-  }
-}
-
-void LynxRobotDriver::DefineDriver()
-{
-  driver_ = std::make_shared<RoboteqDriver>(
-    canopen_manager_.GetMaster(), canopen_settings_.driver_can_ids.at(LynxDriverNames::DEFAULT),
+  auto driver = std::make_shared<RoboteqDriver>(
+    canopen_manager_.GetMaster(), canopen_settings_.driver_can_ids.at(DriverNames::FRONT),
     canopen_settings_.sdo_operation_timeout_ms);
 
   auto left_motor_driver = std::make_shared<RoboteqMotorDriver>(
-    std::dynamic_pointer_cast<RoboteqDriver>(driver_), LynxMotorChannel::LEFT);
+    std::dynamic_pointer_cast<RoboteqDriver>(driver), MotorChannels::LEFT);
   auto right_motor_driver = std::make_shared<RoboteqMotorDriver>(
-    std::dynamic_pointer_cast<RoboteqDriver>(driver_), LynxMotorChannel::RIGHT);
+    std::dynamic_pointer_cast<RoboteqDriver>(driver), MotorChannels::RIGHT);
 
-  driver_->AddMotorDriver(LynxMotorNames::LEFT, left_motor_driver);
-  driver_->AddMotorDriver(LynxMotorNames::RIGHT, right_motor_driver);
-}
+  driver->AddMotorDriver(MotorNames::LEFT, left_motor_driver);
+  driver->AddMotorDriver(MotorNames::RIGHT, right_motor_driver);
 
-void LynxRobotDriver::SetMotorsStates(
-  RoboteqData & data, const MotorDriverState & left_state, const MotorDriverState & right_state,
-  const timespec & current_time)
-{
-  // TODO figure out both motors timestamps
-  bool data_timed_out =
-    (lely::util::from_timespec(current_time) - lely::util::from_timespec(left_state.pos_timestamp) >
-     pdo_motor_states_timeout_ms_) ||
-    (lely::util::from_timespec(current_time) -
-       lely::util::from_timespec(left_state.vel_current_timestamp) >
-     pdo_motor_states_timeout_ms_);
-
-  // Channel 1 - right, Channel 2 - left
-  data.SetMotorsStates(right_state, left_state, data_timed_out);
-}
-
-void LynxRobotDriver::SetDriverState(
-  RoboteqData & data, const DriverState & state, const timespec & current_time)
-{
-  bool data_timed_out = (lely::util::from_timespec(current_time) -
-                           lely::util::from_timespec(state.flags_current_timestamp) >
-                         pdo_driver_state_timeout_ms_) ||
-                        (lely::util::from_timespec(current_time) -
-                           lely::util::from_timespec(state.voltages_temps_timestamp) >
-                         pdo_driver_state_timeout_ms_);
-
-  data.SetDriverState(state, data_timed_out);
+  drivers_.emplace(DriverNames::DEFAULT, driver);
 }
 
 }  // namespace panther_hardware_interfaces
